+++ yuyo/list_status.py
@@ -1001,6 +1001,7 @@ class DiscordBotListService:
         ) as response:
             if shard_id is None:
                 await _log_response("Discordbotlist.com", response, is_global=False)
+                return None  # MyPy compatibility
 
             if response.status in _RETRY_ERROR_CODES:
                 if retry_after := response.headers.get(_RETRY_AFTER_KEY):